update session get running

+++ src/main/java/application/model/Run.java
@@ -10,7 +10,7 @@ public class Run {
     private long id;
     private long athleteId;
     private double distance;
-    private long pace;
+    private double pace;
     private String date;
     private long movingTime;
 
@@ -30,11 +30,11 @@ public class Run {
         this.distance = distance;
     }
 
-    public long getPace() {
+    public double getPace() {
         return pace;
     }
 
-    public void setPace(long pace) {
+    public void setPace(double pace) {
         this.pace = pace;
     }
